Added Login and Registration mechanism.
Code needs to be separated, as in put the registration in the EJB project.

Mail needs to be send to the one that registered.

The index page does not contain any logic, so when you log in you won't see that you actually logged in. The user is set in the session though, so under water he is logged in.

Added little styling